refactor(core): centralize concatenation analysis
